Loot
Finished loot-box script (opening lock and top).
