--- The Dust/Assets/Scripts/Loots/LootController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b86a77ff7206bbf43bceffd2580f8461
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
