+++ The Dust/Assets/Scripts/Storage.cs
@@ -4,6 +4,14 @@ using UnityEngine;
 
 public class Storage : MonoBehaviour
 {
+    enum LootTypes
+    {
+        Aim,
+        Ammo,
+        Armor,
+        Key,
+        Weapon
+    }
 
     // Colliders, that are triggers. For example, they are can not get damaged.
     private List<int> _triggerColliderList = new List<int>();
